+++ change_log.md
@@ -1,3 +1,8 @@
+# 0.3.9 (github, pypi and conda: 2020.12.09)
+* Implemented reading dataframe rownames as pandas index
+* Implementing reading matrices, arrays, tables, solves #5.
+* librdata source updated to commit 8623be2c626392028cd75c5fc45a90e1f70d97b9
+
 # 0.3.8 (github, pypi and conda: 2020.11.15)
 * Fixed sharing flags for opening files on windows